Add BF16 support for all models and fix compatibility issues
- Enable BF16 precision for GPT, BERT, Llama, Mamba, Hyena, T5, Based models
- Auto-detect GPU compute capability ≥ 8.0 for BF16
- Fix T5 model config_kwargs merge issue
- Fix Based model to use standard GPT2 (remove custom library dependency)
- Fix Llama model to create config directly (avoid Meta's gated repo)
- Make Flash Attention 2 optional (requires separate installation)

+++ hf_ehr/models/mamba.py
@@ -19,7 +19,7 @@ class MambaLanguageModel(BaseModel):
         if torch.cuda.get_device_capability('cuda')[0] >= 8:
             print("!!!! USING CACHE !!!!")
             kwargs = {
-                'torch_dtype': torch.float16,
+                'torch_dtype': torch.bfloat16,  # Use BF16 for better training stability
             }
         else:
             kwargs = {}
